fix(dashboard): show B/T units for large token counts, unify number formatting
Token totals past a billion rendered as e.g. 20408.9M. Promote the adaptive
T/B/M/K formatter (magnitude-aware precision, trailing zeros trimmed) from
TokenHeatmap into chartStyles.formatNumber and reuse it across DashboardPage
stat cards, ServiceStatsTable, RequestsView, TokenHistoryChart axes/tooltips,
and QuotaBarItem, replacing six drifting M-capped local copies.
